make sure all compilation paths of cuda/stdpar/cpu/gpu/multicore/gpu-rng/cpu-rng have all symbols
